+++ src/main/java/no/bols/w1/genes/Gene.java
@@ -3,7 +3,7 @@ package no.bols.w1.genes;//
 
 public abstract class Gene {
 
-    public static float MUTATION_CHANCE = 0.4f;
+    public static double MUTATION_CHANCE = 0.4;
 
     public abstract Gene breed(Gene other);
